requirements.txt
# requirements.txt
streamlit
tweepy
python-dotenv